requirements.txt
streamlit
plotly
pandas
pymongo
python-dotenv
apscheduler
lxml
requests
beautifulsoup4 